c324: extract RateDistortion + MutualInformation family from api/client.ts to api/rate-distortion.ts (#441 P1 2.4)
api/client.ts 1095 LOC -> 1059 LOC (-36 LOC). New file: api/rate-distortion.ts (70 LOC).
Moved RateDistortionCurvePoint + RateDistortionCurve + MutualInformationMethod +
MutualInformation types and the 2 runtime methods. client.ts re-exports types
for back-compat; api.rateDistortionCurve / api.mutualInformation delegate
to rateDistortion.* under the hood.
typecheck clean, 44/44 tests pass.
